+++ sphinx/source/changelog.rst
@@ -18,8 +18,8 @@ Changelog
 * The output snapshot argument of :ref:`tap2sna.py` is now optional
 * Added support to :ref:`tap2sna.py <tap2sna-conf>` for register pairs
   (``r[bc]``, ``r[de]`` etc.) in the ``TraceLine`` configuration parameter
-* Added the ``ernieware`` tape-sampling loop
-  :ref:`accelerator <tap2sna-accelerators>`
+* Added the ``ernieware`` and ``housenka`` tape-sampling loop
+  :ref:`accelerators <tap2sna-accelerators>`
 * Added support to :ref:`trace.py` for executing machine code in 128K snapshots
 * Added support to :ref:`trace.py <trace-conf>` for reading configuration
   from `skoolkit.ini`